+++ src/filter.rs
@@ -1,10 +1,10 @@
-pub fn quantize<I: Iterator<Item = f64> + Clone>(samples: I) -> impl Iterator<Item = i16> + Clone {
-    const MAX: f64 = i16::max_value() as f64;
+pub fn quantize<I: Iterator<Item = f32> + Clone>(samples: I) -> impl Iterator<Item = i16> + Clone {
+    const MAX: f32 = i16::max_value() as f32;
     samples.into_iter().map(|sample| (sample * MAX) as i16)
 }
 
-pub fn normalize<I: Iterator<Item = f64> + Clone>(samples: I) -> impl Iterator<Item = f64> + Clone {
-    let normalized: Vec<f64> = samples.into_iter().collect();
+pub fn normalize<I: Iterator<Item = f32> + Clone>(samples: I) -> impl Iterator<Item = f32> + Clone {
+    let normalized: Vec<f32> = samples.into_iter().collect();
 
     let peak = *normalized
         .iter()
